+++ lib/stoplight/light.rb
@@ -131,7 +131,8 @@ module Stoplight
       if error_allowed?(error)
         Stoplight.data_store.clear_failures(name)
       else
-        Stoplight.data_store.record_failure(name, Failure.new(error))
+        failure = Failure.new(error.class.name, error.message)
+        Stoplight.data_store.record_failure(name, failure)
       end
     end