+++ spec/stoplight/failure_spec.rb
@@ -3,9 +3,8 @@
 require 'spec_helper'
 
 describe Stoplight::Failure do
-  subject(:failure) { described_class.new(error, time) }
-  let(:error) { error_class.new(error_message) }
-  let(:error_class) { StandardError }
+  subject(:failure) { described_class.new(error_class, error_message, time) }
+  let(:error_class) { SecureRandom.hex }
   let(:error_message) { SecureRandom.hex }
   let(:time) { Time.now }
 
@@ -13,15 +12,35 @@ describe Stoplight::Failure do
     subject(:result) { described_class.from_json(json) }
     let(:json) { failure.to_json }
 
-    it do
-      expect(result.error).to eq(failure.error)
+    it 'can be round-tripped' do
+      expect(result.error_class).to eq(failure.error_class)
+      expect(result.error_message).to eq(failure.error_message)
       expect(result.time).to be_within(1).of(failure.time)
+    end
+
+    context 'with invalid JSON' do
+      let(:json) { nil }
+
+      it 'does not raise an error' do
+        expect { result }.to_not raise_error
+      end
+
+      it 'returns a self-describing invalid failure' do
+        expect(result.error_class).to eq('Stoplight::Error::InvalidFailure')
+        expect(result.error_message)
+          .to eq('no implicit conversion of nil into String')
+        expect(result.time).to be_within(1).of(Time.now)
+      end
     end
   end
 
   describe '#initialize' do
-    it 'sets the error' do
-      expect(failure.error).to eql(error)
+    it 'sets the error class' do
+      expect(failure.error_class).to eql(error_class)
+    end
+
+    it 'sets the error message' do
+      expect(failure.error_message).to eql(error_message)
     end
 
     it 'sets the time' do
@@ -42,7 +61,8 @@ describe Stoplight::Failure do
     let(:data) { JSON.load(json) }
 
     it 'converts to JSON' do
-      expect(data['error']).to eql(error.inspect)
+      expect(data['error']['class']).to eql(error_class)
+      expect(data['error']['message']).to eql(error_message)
       expect(data['time']).to eql(time.inspect)
     end
   end